Add code to ToggleVisibility Script
I realized that when the Start Menu is open and you open another program it would of course stay open unless you pressed the windows button again.

Of course that is not how windows works, so I added some code to disable objects when a new one is enabled. If we so want to that is.